Remove ChangeLogRunner; Add generate command test
