+++ Tests/DiffFormatterAppTests/OutputTests.swift
@@ -13,11 +13,10 @@ import XCTest
 
 final class OutputTests: XCTestCase {
     func testDefaultOutput() {
-        let outputGenerator: OutputGenerator = .init(
-            configuration: .mock,
-            rawGitLog: defaultInputMock,
-            version: "6.13.0",
-            releaseManager: Configuration.mock.contributors.first
+        let outputGenerator: OutputGenerator = try! .init(
+            options: options(gitLog: defaultInputMock),
+            app: .mock(),
+            env: [:]
         )
 
         assertSnapshot(
@@ -27,11 +26,10 @@ final class OutputTests: XCTestCase {
     }
 
     func testCherryPickedSectionOutput() {
-        let outputGenerator: OutputGenerator = .init(
-            configuration: .mock,
-            rawGitLog: cherryPickedInputMock,
-            version: "6.13.0",
-            releaseManager: Configuration.mock.contributors.first
+        let outputGenerator: OutputGenerator = try! .init(
+            options: options(gitLog: cherryPickedInputMock),
+            app: .mock(),
+            env: [:]
         )
 
         assertSnapshot(
@@ -41,11 +39,10 @@ final class OutputTests: XCTestCase {
     }
 
     func testExcludedSectionOutput() {
-        let outputGenerator: OutputGenerator = .init(
-            configuration: .mockExcludedSection,
-            rawGitLog: defaultInputMock,
-            version: "6.13.0",
-            releaseManager: Configuration.mock.contributors.first
+        let outputGenerator: OutputGenerator = try! .init(
+            options: options(gitLog: cherryPickedInputMock),
+            app: .mock(configuration: .mockExcludedSection),
+            env: [:]
         )
 
         assertSnapshot(
@@ -59,16 +56,17 @@ final class OutputTests: XCTestCase {
 
         let fileManagerMock = FileManagerMock(customConfigPath: customPath)
 
-        let outputGenerator: OutputGenerator = .init(
-            configuration: Configurator(
-                options: .blank,
-                meta: .mock,
-                environment: ["DIFFFORMATTER_CONFIG": customPath],
-                fileManager: fileManagerMock
-                ).configuration,
-            rawGitLog: defaultInputMock,
-            version: "6.13.0",
-            releaseManager: Configuration.mock.contributors.first
+        let configuration = Configurator(
+            options: .blank,
+            meta: .mock,
+            environment: ["DIFFFORMATTER_CONFIG": customPath],
+            fileManager: fileManagerMock
+        ).configuration
+
+        let outputGenerator: OutputGenerator = try! .init(
+            options: options(gitLog: defaultInputMock),
+            app: .mock(configuration: configuration),
+            env: [:]
         )
 
         assertSnapshot(
@@ -100,4 +98,27 @@ final class OutputTests: XCTestCase {
             as: .dump
         )
     }
+
+    private func options(gitLog: String) -> GenerateCommand.Options {
+        return .init(
+            versions: (.init(0, 0, 1), .init(6, 13, 0)),
+            buildNumber: nil,
+            gitLog: gitLog,
+            noFetch: true,
+            noShowVersion: false,
+            releaseManager: Configuration.mock.contributors.first?.email,
+            toPasteBoard: false
+        )
+    }
 }
+
+extension App {
+    static func mock(configuration: Configuration = .mock) -> App {
+        return .init(
+            configuration: configuration,
+            meta: .mock,
+            options: .mock
+        )
+    }
+}
+
